openvswitch: remove ovs-controller package
According to the NEWS message from the main OVS site:
    http://openvswitch.org/releases/NEWS-2.3.0

Excerpt:
 - ovs-controller has been renamed test-controller.  It is no longer
   packaged or installed by default, because too many users assumed
   incorrectly that ovs-controller was a necessary or desirable part
   of an Open vSwitch deployment.

Current feed was updated based on a feed that packaged OVS 1.9,
and subsequently updated.

For now, we'll remove it, since it's not required.
When I tested the feed, I tested it on a setup that has been configured
some time ago, and does not have that package enabled.

So, I'll remove it to prevent confusion.

Signed-off-by: Alexandru Ardelean ardeleanalex@gmail.com