cyrus-sasl: fix typo in install section
An overly specific glob pattern in the package install sections prevents
the unversioned .so symlinks from getting copied into the .ipk.

This commit changes the pattern from xxx.so.* to xxx.so* in order to copy
those symlinks too. Fixes #382.

Also bump the copyright year in the Makefile while we're at it.

Signed-off-by: Jo-Philipp Wich <jow@openwrt.org>