knot: Split utils subpackage
knot-utils consists of kdig, khost, knsec3hash and knsupdate. In order
to reduce the size of software installed on a device it is useful to
include a minimal collection of software and thus it should be possible
to install the utilities individually. bind also creates a subpackage
for each utility.

Signed-off-by: Matthias-Christian Ott <ott@mirix.org>