openvswitch: add +kmod-vxlan dependency for kernels >= 3.12
When OVS detects a kernel version >= 3.12 it does not build
it's own vxlan module and tries to use the kernel's, when building
the OVS kernel module.

I also pushed a patch to the OpenWRT trunk to add a +kmod-vxlan package.

This will add the kernel's vxlan.ko kernel module if it exists.
So, for kernel >= 3.12, this package should exist and be installed
when installing OVS.

Tested on OpenWRT trunk with kernel 3.14.18.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>