protobuf-c: use generic autoreconf fixup
Invoke the generic autoreconf fixup instead of calling the shipped autogen.sh.

This ensures that proper variants of libtoolize, autoconf, automake etc. are
used, otherwise it is not possible to rebuild protobuf-c in the SDK env.

The change requires backport to BB as it currently blocks the rebuild of ocerv.

Signed-off-by: Jo-Philipp Wich <jow@openwrt.org>