python3: revert puting the encodings folder into a python3-codecs
Some notes about the 'encodings' module, which is about 1.7 MB.
Unfortunately that one cannot be moved into the 'python3-codecs'
package, because Python tries to load up all available encodings
at startup.
Some efforts to add a dummy folder/python file have failed so far,
since there's a C code (Python/codecs.c) that tries to evaluate that
all encodings (in the encodings folder/module) are valid.

Basically the encodings module is a repository of encodings,
and it seemst there are quite a few of them.

Maybe a request to upstream Python would work for this, to
make encodings a bit more decoupled from the interpreter.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>