mmc-utils: restore FORTIFY_SOURCE=2
The previous commit
d87f533 mmc-utils: don't set -D_FORTIFY_SOURCE in Makefile
applied the global value of FORITFY_SOURCE also to mmc-utils.
As the package was meant to be built with FORTIFY_SOURCE upstream,
restore that behaviour and make it play nice with the global setting
by undefining the macro before defining it.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>