php5-pecl-libevent: add a patch to force usage of libevent2 headers (refs #1322)
Builtbots currently fail to compile this package with (example message):
 ... /libevent-0.1.0/libevent.c:1403:42: error: 'EVBUFFER_READ' undeclared (first use in this function)
  REGISTER_LONG_CONSTANT("EVBUFFER_READ", EVBUFFER_READ, CONST_CS | CONST_PERSISTENT);

With libevent2, the main include file event.h is deprecated, but includes
enough stuff and other compat includes that it _should work_. I could not figure
out why it fails (all locally builds work out-of-the-box), so this is a trial to
circumvent this compat header file and include the libevent2 headers directly.

While at, add a missing dependency to the sockets extension.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>