[SQM] Fix changes relating to cake lnk layer adjustments
The last batch of changes tried to teach the GUI to pass link layer
options to cake but forgot to actually call the function that parses
the GUI variables and used it as a string insteead. So this fixes that
it also tries to allow the use of the tc_stab link layer adjustment
method with cake so the implementations can be validated against each other
easily. Needs testing...

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>