iotivity: add IoTivity
IoTivity is a Internet of Things framework implementing the Open
Interconnect Consortium Specification.

The current version of IoTivity is still in heavy development and does
not support all its intended features, but I still want to add it to
the packages feed to make it easier for others to extend the OpenWrt
support.

Signed-off-by: Hauke Mehrtens <hauke.mehrtens@lantiq.com>