node: fix build on x86 & x86.64 targets
For x86 and x86_64, nodejs has some special CPU
code that needs to be selected by specifying
the correct CPU name (correct for nodejs).

On OpenWRT x86 is i386 ; node wants ia32 for this.
And x86_64 is x64 on nodejs.

So, we just need to do the proper substitutions.

Note: the ARCH env-var is obtained from CONFIG_ARCH, after
some subtitutions are applied.
So, it shouldn't affect other target archs.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>