keepalived: Fix track_* option in vrrp_instance defn
Also fix ordering of config stanzas

We were parsing the track_script and track_interface definitions to
include the weight param when configuring a vrrp_instance. This is not
correct, as the weight param inside a vrrp instance is used to augment
the one defined in the script.

We were also not taking into account vrrp_script stanzas

This commit skips the parsing and simply lists the name of the
track/vrrp object

Signed-off-by: Ben Kelly <ben@benjii.net>
Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>