php7-pecl-http: fix build when host has php installed (refs #3149)
When php is installed on the host system, then the configure tries
to obtain from this binary whether some extensions are available.
This prevents the cross build when the desired extensions are not
present - which is the usual case.

So add a patch to prevent configure from doing this - we give
this information via environment variable during build already.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>