lcdringer: add new package
Lcdringer is a tool which runs on a Raspberry Pi with an LCD
display. Lcdringer connects to an XMPP server, listens for messages
sent to a particular Jabber ID, and displays these incoming messages
while playing an audible alarm. Lcdringer also responds to the messages
it receives with an indication of whether or not the audible alarm was
acknowledged with a button press.

Signed-off-by: W. Michael Petullo <mike@flyn.org>