znc: add advanced playback module
The advanced playback module makes it possible for IRC clients to avoid
undesired repetitive buffer playback. IRC clients may request the module
to send a partial buffer playback starting from and ending to a certain
point of time.

Particularly useful with (supporting) mobile clients such as Mutter,
Colloquy & others.

Signed-off-by: Kevin Darbyshire-Bryant <kevin@darbyshire-bryant.me.uk>

---

Unfortunately playback.cpp is not included as part of the standard ZNC
modules and so has been added as a patch.