utils/zip: fix LARGE_FILE_SUPPORT in configure script
Package zip is currently broken with a runtime error.
When zip is run, it immediately exits with
"zip warning : Not supported (uzoff_t not same size as zoff_t)"

The issue boils down to the package's configure script which tries to
determine LARGE_FILE_SUPPORT on *host* side. The conftest.c is
compiled and ran on the building host to see whether LFS is given or not.
This will fail when cross-compiling. The patch here is created by Romain Naour,
taken from http://lists.busybox.net/pipermail/buildroot/2015-January/117909.html
Reworked and tested by me. Now LFS is detected and the built binaries work
on the target.

Signed-off-by: Paul Wassi <p.wassi@gmx.at>