keepalived: upgrade to version 1.2.19
Dropped patches, unrequired.
There is a newer version released, but I cannot vouch for it yet.
We've been using this one for about a year now.

Since, I only recently became maintainer of `keepalived`, I will
push this as the current stable one, and start using a newer
version internally, before releasing it to the public.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>