syslog-ng: fix various install problems such as missing plugins, etc
Install the plugins as various functionality is no longer in the
utility itself but is packaged as a .so plugin instead. Disable
plugins with too many dependencies (or too exotic).

Bump the version number on the config file to agree with the package's
version number.

Clean up any .la files from libtool.

Start much earlier so that logging doesn't miss startup messages from
other services.

Signed-off-by: Philip Prindeville <philipp@redfish-solutions.com>