coreutils: add libcap dependency to selected applets
If libcap happens to be present in the environment, coreutils will pick it up
and link some applets against it.

Since the idea of coreutils is to provide a full featured alternative to the
busybox applets, do not inhibit the optional dependency but explicitely
require libcap instead.

Fixes the following error spotted on the buildbots:

    Package coreutils-dir is missing dependencies for the following libraries:
    libcap.so.2

Signed-off-by: Jo-Philipp Wich <jo@mein.io>