netatalk: fix config files permissions
After checking in the ipkg-install dir of netatalk,
it seems that the permissions it sets to the conf-files
are 644.

 # ls -la lede/build_dir/target-mips_24kc_musl/netatalk-3.1.11/ipkg-install/etc
 -rw-r--r-- 1 sandu sandu   226 iun  5 20:53 afp.conf
 -rw-r--r-- 1 sandu sandu  1948 iun  5 20:53 dbus-session.conf
 -rw-r--r-- 1 sandu sandu 25037 iun  5 20:53 extmap.conf

While the Package/netatalk/install build rule overrides
them to 600.

According to
* https://github.com/openwrt/packages/issues/4318
* https://forum.lede-project.org/t/help-with-apple-filesharing-protocol-for-time-machine/3259/5
this is a problem in some setups.

So, this patch changes them to what the netatalk
package creators intended initially (i.e. 644).

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>