net/ratechecker: Add new package
Ratechecker is a program that you can run in your ad-hoc wifi network to
"thoroughly" measure the quality of the links of your network. (By default)
each node sends a series of packets in all possible modulation rates. The
program also acts as the receiver simultaneously and log the packets
received or lost as a bitmap.

Signed-off-by: Sven Eckelmann <sven@narfation.org>