wget: try to fix libunistring dependency
wget 1.91.1 configure script tries to link libunistring
if that is found. That lib is rather large and is not that
essential for normal operations, so instead of depending
on that, try to avoid linking it.

Revert an upstream commit that removed the clearance of
libunistring data in case IRI is explicitly disabled
(like we have it).

Reference to upstream commits:
http://git.savannah.gnu.org/cgit/wget.git/commit/?id=00ae9b4ee2caa6640776e9777e7ef5a374d082c0
http://git.savannah.gnu.org/cgit/wget.git/commit/?id=990f3d67d340d226bbd4187f5d310a4d2bf2bfe4

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>