fwknop: Add start-up dependency on network interface for fwknopd.
Added new "network" section with option "network", which takes network
interface name.

The start-up is migrated to use procd and depend either on the "network"
interface (after resolving it to a physical device), or on the PCAP_INTF
option from "config" section (usual place for raw interface name for
fwknopd). When the uci_enabled option is disabled, the value of PCAP_INTF
is taken from the user-provided fwknopd.conf.

Also fixed UCI_ENABLED variable evaluation.

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>