net/keepalived: add quoting for print_elems_indent function
If you add a notify_* script which is not only a path to a script or
to a bin file, it is necessary to quote to whole string. If you do not
quote this the config will not get accepted by keepalived and so will
not start. This will fix this issue.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>