net: mstpd: new package (multiple spanning tree daemon)
This change adds support for mstpd (Multiple Spanning Tree Protocol
Daemon).

mstpd works reasonably well with RSTP.
MSTP protocol works ok, but is known to have some issues with some managed
switches.

In order to get this to work, each physical switch port needs to have it's
own software network interface (so, for example: port 0 <==> eth0). This
means that this is suited mostly for higher end devices that can process
STP packets in software.

An interface for `swconfig` or Linux's DSA or switchdev would haven been
interesting, but it never materialized.

Adding this in the OpenWrt packages feed may provide some interest or
feedback on whether `mstpd` should do more, to integrate with managed
switches and offer some basis for Linux (through OpenWrt) as an OS for
managed switches.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>