luaposix: musl compatibility with XOPEN_REALTIME
musl-libc doesn't define _XOPEN_REALTIME as it doesn't implement the
full set of the realtime operations.  However, it _does_ implement
POSIX_TIMERS, which is what luaposix _should_ be checking for in the
posix.time module.

I've filed https://github.com/luaposix/luaposix/issues/295 to track this
upstream, but this is a simpler, verified fix for our older version of
luaposix.

This restores functions like posix.time.clock_gettime(CLOCK_MONOTONIC)
for instance, which was available with pre-musl builds of luaposix.

Signed-off-by: Karl Palsson <karlp@etactica.com>