transmission: Add rpc_host_whitelist support.
It was requested as it can be useful under certain circumstances.

Disabled rpc_whitelist by default. Not only is there a firewall, but it denies access when IP address of the device is changed.

Added group support in UCI. Fixes cases where group does not match the user (nobody:nogroup).

Signed-off-by: Rosen Penev <rosenp@gmail.com>