tvheadend: minor bugfix version bump to 4.0.10
Update the package to the last release of the old 4.0 series.

Note that this version is still from the obsolete 4.0 series,
but updates the source base a bit.

This also enables dropping one of the GCC 7.3 fixes.

Compiled and briefly run-tested with ipq806x and mvebu,
but no actual operations done.

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>