python,python3: remove `--ignore-installed` flag for host packages
This was copied over from python-packages, when support for installing
packages host-side (via pip) was added.

Based on the discussion on this commit:
  https://github.com/openwrt/packages/commit/b125664de72166223ec86e22cdc8a3584cd2d3ff
it was mentioned that removing this may add more benefit in terms of
reducing build time, because packages won't get reinstalled every time.

I'm not entirely sure about any potential side-effects of this, but it's
worth trying it out.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>