keepalived: disable libnl
When libnl is enabled and libnl-genl is disabled, libnl-genl-3.so.200 is
installed by InstallDev, and keepalived picks it up anyway.
This causes build to fail with the following error:

Package keepalived is missing dependencies for the following libraries:
libnl-genl-3.so.200

Instead of having a conditional dependency it is better to have a
consistent build. Disable libnl to enforce this.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>