radicale2: Update CalDAV/CardDAV server to v2.x
Radicale 2.x adds support for many new clients,
bug-fixes, etc so add v2 of this application.

We do it as a separate package for those not
ready to switch (it's not an straight inplace
upgrade from 1.x).

We do however CONFLICT with 1.x as they can't
be run side-by-side on the same host (without
containers for somesuch).

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>