prometheus-statsd-exporter: switch config to openwrt interface names
Drop the config knobs '*_address' and introduce '*_interface'
and '*_ipv6' instead.

'*_interface' takes an openwrt interface name ('loopback', 'lan',
'wan' etc), from which the primary IP is used to listen on. If
the matching '*_ipv6' is set to '1', the IPv6 adress will be used,
IPv4 elsewise.

procd interface triggers are now combined with this, so if a listen
interface is not yet configured when the init script is executed, the
process start is defered, and the trigger takes care of that once
the interfaces are ready.

Signed-off-by: Andre Heider <a.heider@gmail.com>