graphicsmagick: new package, an imagemagick alternative
GraphicsMagick is a fork of ImageMagick, licensed under the MIT license.
It requires less space and it is [reportedly] faster than ImageMagick.

GraphicsMagick's installation footprint is:

x86_64: 4.3 MB, and
brcm47xx-mips74k (mipsel_74kc): 3.7 MB.

The shared libraries occupy 2.2 MB (mipsel_74kc). The 90 GraphicsMagick's
modules occupy 2.5 MB. It may be possible to reduce the installation
footprint by introducing build parameters to control the selection of
modules. In view of the large number of modules and the possibility of
breakage due to module interdependencies or other reasons, such attempt
is not made at this time.

Signed-off-by: Val Kulkov <val.kulkov@gmail.com>