postgresql: Revert adding build dependency to zlib/host
Revert the addition of build dependency in commit f0b73ca3b
to a non-existent host build of zlib.

The host build of zlib was removed already in April 2018 by
https://github.com/openwrt/openwrt/commit/8dcd941d8b934891676a8d4bbef1ee78e89a4bf7#diff-1ed408c61d79f9c6c5d197333e94ce8d
which made zlib a build tool defined in /tools

The newly introduced build dependency causes always a warning like:
   WARNING: Makefile 'package/feeds/packages/postgresql/Makefile'
   has a build dependency on 'zlib/host', which does not exist

Not sure what was the error that f0b73ca3b tried to fix,
but reference to a non-existent host build is not the solution.

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>