isc-dhcp: Remove --disable-tracing from configure args
This fixes a crash that happens when dhcpd is configured with a failover
peer, and the failover peer goes down. The crash is due to a dereference
of a freed object. When tracing is enabled (which is the default) the
object in question is referenced by the tracing code and so doesn't get
freed prematurely. I have observed this crash on two different target
platforms (mips and x86_64), and it is reproducible on non-OpenWRT
distros by building isc-dhcpd using --disable-tracing.

This has been reported to ISC, but their response was that it's a low
priority as the simple work-around is to leave tracing enabled.

Re-enabling the tracing code only increases the size of the executable
by about 24KB.

Signed-off-by: Heath Kehoe <yaheath@gmail.com>